Separate API-related files
