+++ app/zosia16/urls.py
@@ -13,23 +13,27 @@ Including another URLconf
     1. Import the include() function: from django.conf.urls import url, include
     2. Add a URL to urlpatterns:  url(r'^blog/', include('blog.urls'))
 """
-from django.conf.urls import url, include
 from django.conf import settings
-from django.contrib import admin
+from django.conf.urls import include, url
 from django.conf.urls.static import static
+from django.contrib import admin
 
-urlpatterns = [
-    url(r'^admin/', admin.site.urls),
-    url(r'^accounts/', include('users.urls')),
-    url(r'', include('conferences.urls')),
-    url(r'^rooms/', include('rooms.urls')),
-    url(r'^blog/', include('blog.urls')),
-    url(r'^sponsors/', include('sponsors.urls')),
-    url(r'^lectures/', include('lectures.urls')),
-    url(r'^questions/', include('questions.urls')),
-    # NOTE: It only serve static files when debug=True
-] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+urlpatterns = \
+    [
+        url(r'^admin/', admin.site.urls),
+        url(r'^accounts/', include('users.urls')),
+        url(r'', include('conferences.urls')),
+        url(r'^rooms/', include('rooms.urls')),
+        url(r'^blog/', include('blog.urls')),
+        url(r'^sponsors/', include('sponsors.urls')),
+        url(r'^lectures/', include('lectures.urls')),
+        url(r'^questions/', include('questions.urls')),
+
+        url(r'^api/rooms/', include('rooms.api.urls')),
+    ] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+# NOTE: It only serve static files when debug=True
 
 if settings.DEBUG:
     import debug_toolbar
+
     urlpatterns += [url(r'^__debug__/', include(debug_toolbar.urls))]
